sw: Factor out the connection functionality to RouterClient
